Created complexObjects test file
